# Implement off-heap quantized scoring