--- a/lucene/core/src/java/org/apache/lucene/internal/vectorization/VectorUtilSupport.java
+++ b/lucene/core/src/java/org/apache/lucene/internal/vectorization/VectorUtilSupport.java
@@ -36,17 +36,39 @@ public interface VectorUtilSupport {
   /** Returns the dot product computed over signed bytes. */
   int dotProduct(byte[] a, byte[] b);
 
+  /** Returns the dot product computed over unsigned half-bytes, both uncompressed. */
+  int int4DotProduct(byte[] a, byte[] b);
+
+  /** Returns the dot product computed over unsigned half-bytes, one compressed. */
+  int int4DotProductSinglePacked(byte[] unpacked, byte[] packed);
+
+  /** Returns the dot product computed over unsigned half-bytes, both compressed. */
+  int int4DotProductBothPacked(byte[] a, byte[] b);
+
   /** Returns the dot product computed as though the bytes were unsigned. */
   int uint8DotProduct(byte[] a, byte[] b);
-
-  /** Returns the dot product over the computed bytes, assuming the values are int4 encoded. */
-  int int4DotProduct(byte[] a, boolean apacked, byte[] b, boolean bpacked);
 
   /** Returns the cosine similarity between the two byte vectors. */
   float cosine(byte[] a, byte[] b);
 
   /** Returns the sum of squared differences of the two byte vectors. */
   int squareDistance(byte[] a, byte[] b);
+
+  /**
+   * Returns the sum of squared differences between two unsigned half-byte vectors, both
+   * uncompressed.
+   */
+  int int4SquareDistance(byte[] a, byte[] b);
+
+  /**
+   * Returns the sum of squared differences between two unsigned half-byte vectors, one compressed.
+   */
+  int int4SquareDistanceSinglePacked(byte[] unpacked, byte[] packed);
+
+  /**
+   * Returns the sum of squared differences between two unsigned half-byte vectors, both compressed.
+   */
+  int int4SquareDistanceBothPacked(byte[] a, byte[] b);
 
   /** Returns the sum of squared differences of the two unsigned byte vectors. */
   int uint8SquareDistance(byte[] a, byte[] b);