--- a/lucene/core/src/java/org/apache/lucene/util/VectorUtil.java
+++ b/lucene/core/src/java/org/apache/lucene/util/VectorUtil.java
@@ -113,6 +113,37 @@ public final class VectorUtil {
     return IMPL.squareDistance(a, b);
   }
 
+  /** Returns the sum of squared differences between two uint4 (values between [0,15]) vectors. */
+  public static int int4SquareDistance(byte[] a, byte[] b) {
+    if (a.length != b.length) {
+      throw new IllegalArgumentException("vector dimensions differ: " + a.length + "!=" + b.length);
+    }
+    return IMPL.int4SquareDistance(a, b);
+  }
+
+  /**
+   * Returns the sum of squared differences between two uint4 (values between [0,15]) vectors. The
+   * second vector is considered "packed" (i.e. every byte representing two values).
+   */
+  public static int int4SquareDistanceSinglePacked(byte[] unpacked, byte[] packed) {
+    if (packed.length != ((unpacked.length + 1) >> 1)) {
+      throw new IllegalArgumentException(
+          "vector dimensions differ: " + unpacked.length + "!= 2 * " + packed.length);
+    }
+    return IMPL.int4SquareDistanceSinglePacked(unpacked, packed);
+  }
+
+  /**
+   * Returns the sum of squared differences between two uint4 (values between [0,15]) vectors. Both
+   * vectors are considered "packed" (i.e. every byte representing two values).
+   */
+  public static int int4SquareDistanceBothPacked(byte[] a, byte[] b) {
+    if (a.length != b.length) {
+      throw new IllegalArgumentException("vector dimensions differ: " + a.length + "!=" + b.length);
+    }
+    return IMPL.int4SquareDistanceBothPacked(a, b);
+  }
+
   /** Returns the sum of squared differences of the two vectors where each byte is unsigned */
   public static int uint8SquareDistance(byte[] a, byte[] b) {
     if (a.length != b.length) {
@@ -189,15 +220,22 @@ public final class VectorUtil {
     return IMPL.uint8DotProduct(a, b);
   }
 
+  /**
+   * Dot product computed over uint4 (values between [0,15]) bytes.
+   *
+   * @param a bytes containing a vector
+   * @param b bytes containing another vector, of the same dimension
+   * @return the value of the dot product of the two vectors
+   */
   public static int int4DotProduct(byte[] a, byte[] b) {
     if (a.length != b.length) {
       throw new IllegalArgumentException("vector dimensions differ: " + a.length + "!=" + b.length);
     }
-    return IMPL.int4DotProduct(a, false, b, false);
+    return IMPL.int4DotProduct(a, b);
   }
 
   /**
-   * Dot product computed over int4 (values between [0,15]) bytes. The second vector is considered
+   * Dot product computed over uint4 (values between [0,15]) bytes. The second vector is considered
    * "packed" (i.e. every byte representing two values). The following packing is assumed:
    *
    * <pre class="prettyprint lang-java">
@@ -211,12 +249,28 @@ public final class VectorUtil {
    * @param packed the packed vector, of length {@code (unpacked.length + 1) / 2}
    * @return the value of the dot product of the two vectors
    */
-  public static int int4DotProductPacked(byte[] unpacked, byte[] packed) {
+  public static int int4DotProductSinglePacked(byte[] unpacked, byte[] packed) {
     if (packed.length != ((unpacked.length + 1) >> 1)) {
       throw new IllegalArgumentException(
           "vector dimensions differ: " + unpacked.length + "!= 2 * " + packed.length);
     }
-    return IMPL.int4DotProduct(unpacked, false, packed, true);
+    return IMPL.int4DotProductSinglePacked(unpacked, packed);
+  }
+
+  /**
+   * Dot product computed over uint4 (values between [0,15]) bytes. Both vectors are considered
+   * "packed" (i.e. every byte representing two values).
+   *
+   * @param a bytes containing a packed vector
+   * @param b bytes containing another packed vector, of the same dimension
+   * @return the value of the dot product of the two vectors
+   */
+  public static int int4DotProductBothPacked(byte[] a, byte[] b) {
+    if (a.length != b.length) {
+      throw new IllegalArgumentException(
+          "vector dimensions differ: " + a.length + " != " + b.length);
+    }
+    return IMPL.int4DotProductBothPacked(a, b);
   }
 
   /**