… lucene/CHANGES.txt
… lucene/benchmark-jmh/src/java/org/apache/lucene/benchmark/jmh/VectorUtilBenchmark.java
… lucene/core/src/java24/org/apache/lucene/internal/vectorization/PanamaVectorUtilSupport.java
… lucene/core/src/test/org/apache/lucene/codecs/lucene99/TestLucene99HnswQuantizedVectorsFormat.java
… lucene/core/src/test/org/apache/lucene/codecs/lucene99/TestLucene99ScalarQuantizedVectorsFormat.java
… lucene/core/src/test/org/apache/lucene/internal/vectorization/TestVectorUtilSupport.java