@ChrisHegarty 2025-06-30
> ..then we see a performance regression:
> ...
> Maybe I'm missing something obvious, but I haven't found the root cause yet..

yeah. I've seen similar before. You might be hitting a problem with the loop bound not being hoisted. I will try to take a look.