@kaivalnp 2025-06-30
Thanks @ChrisHegarty! I saw that we [use](https://github.com/apache/lucene/blob/f4339ee2aea65bdf1efb2a3c196c3e9a4adf9d67/lucene/core/src/java24/org/apache/lucene/internal/vectorization/Lucene99MemorySegmentByteVectorScorer.java#L64) a heap-backed `MemorySegment` while scoring byte vectors -- so I opened #14874 to investigate if we can improve performance by moving to an off-heap query