@kaivalnp 2025-07-15
After [this conversation](https://github.com/apache/lucene/pull/14874#issuecomment-3071054127) I re-ran some benchmarks with `-XX:CompileCommand=inline,*PanamaVectorUtilSupport.*` to force inlining of dot product functions

`main`:

| recall | latency | index time (s) | force merge (s) | quantization | index      |
| ------ | ------- | -------------- | --------------- | ------------ | ---------- |
| 0.530  | 1.675   | 75.65          | 59.97           | 4 bit        | fresh      |
| 0.860  | 1.844   | 81.75          | 71.79           | 7 bit        | fresh      |
| 0.530  | 1.582   | -              | -               | 4 bit        | no reindex |
| 0.860  | 1.859   | -              | -               | 7 bit        | no reindex |
| 0.529  | 1.682   | 79.32          | 62.62           | 4 bit        | reindex    |
| 0.859  | 1.821   | 103.78         | 48.36           | 7 bit        | reindex    |

This PR:

| recall | latency | index time (s) | force merge (s) | quantization | index      | query type                    |
| ------ | ------- | -------------- | --------------- | ------------ | ---------- | ----------------------------- |
| 0.529  | 2.132   | 131.86         | 85.86           | 4 bit        | fresh      | `MemorySegment.ofArray`       |
| 0.858  | 2.797   | 133.71         | 80.81           | 7 bit        | fresh      | `MemorySegment.ofArray`       |
| 0.529  | 2.081   | -              | -               | 4 bit        | no reindex | `MemorySegment.ofArray`       |
| 0.858  | 1.670   | -              | -               | 7 bit        | no reindex | `MemorySegment.ofArray`       |
| 0.529  | 2.140   | 130.82         | 85.08           | 4 bit        | reindex    | `MemorySegment.ofArray`       |
| 0.858  | 2.883   | 132.66         | 83.05           | 7 bit        | reindex    | `MemorySegment.ofArray`       |
| 0.529  | 1.511   | 164.22         | 110.18          | 4 bit        | fresh      | `Arena.ofAuto().allocateFrom` |
| 0.859  | 1.728   | 132.41         | 81.35           | 7 bit        | fresh      | `Arena.ofAuto().allocateFrom` |
| 0.529  | 1.551   | -              | -               | 4 bit        | no reindex | `Arena.ofAuto().allocateFrom` |
| 0.859  | 1.704   | -              | -               | 7 bit        | no reindex | `Arena.ofAuto().allocateFrom` |
| 0.529  | 1.574   | 164.04         | 112.60          | 4 bit        | reindex    | `Arena.ofAuto().allocateFrom` |
| 0.859  | 1.774   | 135.20         | 83.72           | 7 bit        | reindex    | `Arena.ofAuto().allocateFrom` |

Looks like the changes in this PR have a small benefit on the search side, but slow down indexing by a lot..